README 5
Create a student class, that will store the details of the Student
Reg. No.
Name
Email
Phone
Class
Department
The class will have a constructor to initialize the values of the Student and a method to print the Details of the Student.

In the main-method class, create an array of Student Class to hold maximum details of 100 Students.
In the menu-driven program, the menu options will have
Add a student
Adds the details of 1 student to the array of Student
Search for a student
Search for the details of a student from the array of Student
(Optional) Searching can be done with Name & Register Number
Display all students
Displays the details of all students
Note: Make use of functions to implement the menu options

Extension (Optional):
Save the details of each student in a file, with his name with the help of a method in student class, when the student is getting created.